modify: スタンプをstampに切り出し
