+++ app/basic_information/stamp/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class StampConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "basic_information.stamp"
