+++ app/en100_app/settings.py
@@ -28,6 +28,7 @@ INSTALLED_APPS = [
     "basic_information.users",
     "basic_information.questions",
     "basic_information.mypage",
+    "basic_information.stamp",
 ]
 
 MIDDLEWARE = [
